+++ cmsaf/NEWS.md
@@ -4,6 +4,7 @@
 
 - Add new operators timavg, num_above, num_below, num_equal,
   mon_num_above, mon_num_below, mon_num_equal
+- Add operators cmsaf.mul and cmsaf.div to Toolbox
 
 ## 3.1.1